Rename WorkloadResources file
Filename reflected old class name
