+++ valohai_yaml/objs/workload_resources.py
